# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"videos","description":"Video application","contact":{"name":"Agustin Nieto","email":"santiago-n@live.com.ar"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Video application"}],"paths":{"/players":{"post":{"tags":["player-controller"],"summary":"Create a new player","description":"Returns the player created with an id. If the another player exists with the same username or email then will return a 400 status code. Additionally, the email must be valid and the password must have at least 8 characters and contain at least one number, one lowercase letter, one uppercase letter, and one special character.","operationId":"savePlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Player"}}},"required":true},"responses":{"400":{"description":"Username or Email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/matches":{"post":{"tags":["match-controller"],"summary":"Create a new Player","description":"Returns the player created with an id. If the another player exists with the same username or email then will return a 400 status code. Additionally, the email must be valid and the password must have at least 8 characters and contain at least one number, one lowercase letter, one uppercase letter, and one special character.","operationId":"saveMatch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MatchDto"}}},"required":true},"responses":{"400":{"description":"Username or Email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/login":{"post":{"tags":["login-controller"],"summary":"Login a player in the platform","description":"Return the player logged in if the credentials are correct.","operationId":"loginPlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Credential"}}},"required":true},"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"404":{"description":"User is not registered","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/login/v2":{"post":{"tags":["login-controller"],"summary":"Login a player in the platform","description":"Return the player logged in if the credentials are correct.","operationId":"loginPlayer_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CredentialV2"}}},"required":true},"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"400":{"description":"User or Email already used","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/players/{id}":{"get":{"tags":["player-controller"],"summary":"Get a player by id","description":"Returns a player by their id. If the player doesn't exist will then throw a 404 status code","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/players/test/{id}":{"get":{"tags":["player-controller"],"operationId":"getTest","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app is alive will respond with pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/matches/player/{id}/":{"get":{"tags":["match-controller"],"operationId":"getMatchesOfPlayer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Match"}}}}}}}}},"components":{"schemas":{"Player":{"required":["email","password","userName"],"type":"object","properties":{"id":{"title":"Player ID","type":"integer","description":"The Player ID","format":"int64","example":7},"userName":{"title":"Player Username","type":"string","description":"The Players username","example":"MyUsername"},"password":{"title":"Player Password","type":"string","description":"The Players password","example":"Password1337%"},"email":{"title":"Player e-mail","type":"string","description":"The Player email","example":"example@mail.com"},"avatar":{"title":"Player avatar","type":"string","description":"The Player avatar","example":"https://i.pinimg.com/564x/dd/17/a2/dd17a2b3aac1526d8ae0ba6df968ead4.jpg"},"lastLogin":{"title":"Player last login","type":"string","description":"Last time seen online","format":"date-time","nullable":true}}},"MatchDto":{"required":["gameId","playerId"],"type":"object","properties":{"gameId":{"type":"integer","format":"int64"},"playerId":{"type":"integer","format":"int64"}}},"Credential":{"type":"object","properties":{"identity":{"oneOf":[{"$ref":"#/components/schemas/EmailIdentity"},{"$ref":"#/components/schemas/UserNameIdentity"}]},"password":{"type":"string"}}},"EmailIdentity":{"required":["email","identity_type"],"type":"object","allOf":[{"$ref":"#/components/schemas/Identity"},{"type":"object","properties":{"email":{"title":"Email to log in","type":"string","description":"The player's email","example":"example@mail.com"}}}]},"Identity":{"required":["identity_type"],"type":"object","properties":{"identity_type":{"title":"Type of identity used to log in","type":"string","description":"The type of identity provided to log in","example":"USERNAME or EMAIL","enum":["USERNAME","EMAIL"]}},"discriminator":{"propertyName":"identity_type"}},"UserNameIdentity":{"required":["identity_type","user_name"],"type":"object","allOf":[{"$ref":"#/components/schemas/Identity"},{"type":"object","properties":{"user_name":{"title":"Username to log in","type":"string","description":"The player's username","example":"myUserName"}}}]},"CredentialV2":{"required":["identity","password"],"type":"object","properties":{"password":{"type":"string"},"identity":{"title":"Email or Username to log in","type":"string","description":"The player's email or username","example":"example@mail.com OR myUserName"}}},"Game":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"rules":{"type":"string"}}},"Match":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"game":{"$ref":"#/components/schemas/Game"},"player":{"$ref":"#/components/schemas/Player"},"createdDate":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["STARTED","FINISHED","CANCELED"]}}}}}}